rename header files to not have duplicate header file names
