+++ include/client/client_gameobject.hpp
