+++ src/server/gameobject.cpp
@@ -1,4 +1,4 @@
-#include "gameobject.hpp"
+#include "server_gameobject.hpp"
 
 void GameObject::applyMovement(const glm::vec3 &direction) {
   glm::vec3 pos = transform->getPosition();
